# Introduce more faster and memory optimal BitSet as replacment of Set<i32>