--- a/src/util/collections.ts
+++ b/src/util/collections.ts
@@ -24,3 +24,67 @@ export function uniqueMap<K,V>(original: Map<K,V> | null = null, overrides: Map<
   }
   return cloned;
 }
+
+export class BitSet {
+  words!: Uint32Array;
+
+  constructor() {
+    this.clear();
+  }
+
+  get size(): i32 {
+    var count = 0;
+    var words = this.words;
+    for (let i = 0, len = words.length; i < len; i++) {
+      count += popcnt(unchecked(words[i]));
+    }
+    return count;
+  }
+
+  add(index: i32): this {
+    var idx = index >>> 5;
+    if (idx > this.words.length) {
+      this.resize(index);
+    }
+    unchecked(this.words[idx] |= 1 << index);
+    return this;
+  }
+
+  delete(index: i32): void {
+    var idx = index >>> 5;
+    var words = this.words;
+    if (idx >= words.length) return;
+    unchecked(words[idx] &= ~(1 << index));
+  }
+
+  has(index: i32): bool {
+    return (this.words[index >>> 5] & (1 << index)) !== 0;
+  }
+
+  clear(): void {
+    this.words = new Uint32Array(8);
+  }
+
+  private resize(index: i32): void {
+    var newWords = new Uint32Array((index + 64) >>> 5);
+    newWords.set(this.words);
+    this.words = newWords;
+  }
+
+  toArray(): i32[] {
+    var res = new Array<i32>(this.size);
+    for (let i = 0, p = 0, len = this.words.length; i < len; ++i) {
+      let word = unchecked(this.words[i]);
+      while (word) {
+        let mask = word & -word;
+        res[p++] = (i << 5) + popcnt(mask - 1);
+        word ^= mask;
+      }
+    }
+    return res;
+  }
+
+  toString(): string {
+    return `BitSet { ${this.toArray()} }`;
+  }
+}