Benchmark results for 1M items (JS):
```
BitSet add: 7.995ms
BitSet has: 4.894ms

Set<i32> add: 15.898ms
Set<i32> has: 15.414ms
```

Benchmark results for 100M items (Wasm):
```
BitSet add: 39.5000
BitSet has: 29.3000

Set<i32> add: 194.6000
Set<i32> has: 207.3000
```

- [x] I've read the contributing guidelines
- [x] I've added my name and email to the NOTICE file