@dcodeIO 2022-07-06
What I'd find interesting in this context is a comparison of `Set` vs `BitSet` when compiled to Wasm. Mostly just curious, but perhaps the existing benchmark in JS can be extended? :)