@MaxGraey 2022-07-06
Here AS bench BitSet vs Set<i32>:

[playground](https://www.assemblyscript.org/editor.html#IyFvcHRpbWl6ZT1zcGVlZCZydW50aW1lPWluY3JlbWVudGFsCmNsYXNzIEJpdFNldCB7CiAgd29yZHMhOiBVaW50MzJBcnJheTsKCiAgY29uc3RydWN0b3IoKSB7CiAgICB0aGlzLmNsZWFyKCk7CiAgfQoKICBnZXQgc2l6ZSgpOiBpMzIgewogICAgdmFyIGNvdW50ID0gMDsKICAgIHZhciB3b3JkcyA9IHRoaXMud29yZHM7CiAgICBmb3IgKGxldCBpID0gMCwgbGVuID0gd29yZHMubGVuZ3RoOyBpIDwgbGVuOyBpKyspIHsKICAgICAgbGV0IHdvcmQgPSB1bmNoZWNrZWQod29yZHNbaV0pOwogICAgICBpZiAod29yZCkgY291bnQgKz0gcG9wY250KHdvcmQpOwogICAgfQogICAgcmV0dXJuIGNvdW50OwogIH0KCiAgYWRkKGluZGV4OiBpMzIpOiB0aGlzIHsKICAgIHZhciBpZHggPSBpbmRleCA+Pj4gNTsKICAgIHZhciB3b3JkcyA9IHRoaXMud29yZHM7CiAgICBpZiAoaWR4ID49IHdvcmRzLmxlbmd0aCkgeyAvLyByZXNpemUKICAgICAgdGhpcy53b3JkcyA9IG5ldyBVaW50MzJBcnJheShpZHggKyAxNik7CiAgICAgIHRoaXMud29yZHMuc2V0KHdvcmRzKTsKICAgICAgd29yZHMgPSB0aGlzLndvcmRzOwogICAgfQogICAgdW5jaGVja2VkKHdvcmRzW2lkeF0gfD0gMSA8PCBpbmRleCk7CiAgICByZXR1cm4gdGhpczsKICB9CgogIGRlbGV0ZShpbmRleDogaTMyKTogdm9pZCB7CiAgICB2YXIgaWR4ID0gaW5kZXggPj4+IDU7CiAgICB2YXIgd29yZHMgPSB0aGlzLndvcmRzOwogICAgaWYgKGlkeCA+PSB3b3Jkcy5sZW5ndGgpIHJldHVybjsKICAgIHVuY2hlY2tlZCh3b3Jkc1tpZHhdICY9IH4oMSA8PCBpbmRleCkpOwogIH0KCiAgaGFzKGluZGV4OiBpMzIpOiBib29sIHsKICAgIHZhciBpZHggPSBpbmRleCA+Pj4gNTsKICAgIHZhciB3b3JkcyA9IHRoaXMud29yZHM7CiAgICBpZiAoaWR4ID49IHdvcmRzLmxlbmd0aCkgcmV0dXJuIGZhbHNlOwogICAgcmV0dXJuICh1bmNoZWNrZWQod29yZHNbaW5kZXggPj4+IDVdKSAmICgxIDw8IGluZGV4KSkgIT09IDA7CiAgfQoKICBjbGVhcigpOiB2b2lkIHsKICAgIHRoaXMud29yZHMgPSBuZXcgVWludDMyQXJyYXkoMTYpOwogIH0KCiAgdG9BcnJheSgpOiBpMzJbXSB7CiAgICB2YXIgcmVzID0gbmV3IEFycmF5PGkzMj4odGhpcy5zaXplKTsKICAgIGZvciAobGV0IGkgPSAwLCBwID0gMCwgbGVuID0gdGhpcy53b3Jkcy5sZW5ndGg7IGkgPCBsZW47ICsraSkgewogICAgICBsZXQgd29yZCA9IHVuY2hlY2tlZCh0aGlzLndvcmRzW2ldKTsKICAgICAgd2hpbGUgKHdvcmQpIHsKICAgICAgICBsZXQgbWFzayA9IHdvcmQgJiAtd29yZDsKICAgICAgICB1bmNoZWNrZWQocmVzW3ArK10gPSAoaSA8PCA1KSArIHBvcGNudChtYXNrIC0gMSkpOwogICAgICAgIHdvcmQgXj0gbWFzazsKICAgICAgfQogICAgfQogICAgcmV0dXJuIHJlczsKICB9CgogIHRvU3RyaW5nKCk6IHN0cmluZyB7CiAgICByZXR1cm4gYEJpdFNldCB7ICR7dGhpcy50b0FycmF5KCl9IH1gOwogIH0KfQoKY29uc3QgSVRFUlMgICA9IDEwOwpjb25zdCBJVEVSU19JID0gMTBfMDAwXzAwMDsKCmNvbnN0IGJpdHNldCA9IG5ldyBCaXRTZXQoKTsKY29uc3Qgc2V0ID0gbmV3IFNldDxpMzI+KCk7CgpleHBvcnQgZnVuY3Rpb24gYmVuY2hBZGQoKTogdm9pZCB7CiAgZm9yIChsZXQgaiA9IDA7IGogPCBJVEVSUzsgaisrKSB7CiAgICBmb3IgKGxldCBpID0gMDsgaSA8IElURVJTX0kgLyAxMDsgaSsrKSB7CiAgICAgIGJpdHNldC5hZGQoaSk7CiAgICB9CiAgfQp9CgpleHBvcnQgZnVuY3Rpb24gYmVuY2hIYXMoKTogaTMyIHsKICBsZXQgYWNjID0gMDsKICBmb3IgKGxldCBqID0gMDsgaiA8IElURVJTOyBqKyspIHsKICAgIGZvciAobGV0IGkgPSAwOyBpIDwgSVRFUlNfSSAvIDEwOyBpKyspIHsKICAgICAgYWNjIF49IGkzMihiaXRzZXQuaGFzKGkpKTsKICAgIH0KICB9CiAgcmV0dXJuIGFjYzsKfQoKZXhwb3J0IGZ1bmN0aW9uIGNsZWFyKCk6IHZvaWQgewogIGJpdHNldC5jbGVhcigpOwp9CgoKCmV4cG9ydCBmdW5jdGlvbiBiZW5jaFNldEFkZCgpOiB2b2lkIHsKICBmb3IgKGxldCBqID0gMDsgaiA8IElURVJTOyBqKyspIHsKICAgIGZvciAobGV0IGkgPSAwOyBpIDwgSVRFUlNfSSAvIDEwOyBpKyspIHsKICAgICAgc2V0LmFkZChpKTsKICAgIH0KICB9Cn0KCmV4cG9ydCBmdW5jdGlvbiBiZW5jaFNldEhhcygpOiBpMzIgewogIGxldCBhY2MgPSAwOwogIGZvciAobGV0IGogPSAwOyBqIDwgSVRFUlM7IGorKykgewogICAgZm9yIChsZXQgaSA9IDA7IGkgPCBJVEVSU19JIC8gMTA7IGkrKykgewogICAgICBhY2MgXj0gaTMyKHNldC5oYXMoaSkpOwogICAgfQogIH0KICByZXR1cm4gYWNjOwp9CgpleHBvcnQgZnVuY3Rpb24gY2xlYXJTZXQoKTogdm9pZCB7CiAgc2V0LmNsZWFyKCk7Cn0KIyFodG1sCjx0ZXh0YXJlYSBpZD0ib3V0cHV0IiBzdHlsZT0iaGVpZ2h0OiAxMDAlOyB3aWR0aDogMTAwJSIgcmVhZG9ubHk+PC90ZXh0YXJlYT4KPHNjcmlwdCB0eXBlPSJtb2R1bGUiPgpjb25zdCBleHBvcnRzID0gYXdhaXQgaW5zdGFudGlhdGUoYXdhaXQgY29tcGlsZSgpLCB7IC8qIGltcG9ydHMgKi8gfSkKY29uc3Qgb3V0cHV0ID0gZG9jdW1lbnQuZ2V0RWxlbWVudEJ5SWQoJ291dHB1dCcpCmNvbnN0IHsgCiAgYmVuY2hBZGQsIGJlbmNoSGFzLCBjbGVhciwKICBiZW5jaFNldEFkZCwgYmVuY2hTZXRIYXMsIGNsZWFyU2V0Cn0gPSBleHBvcnRzOwoKZnVuY3Rpb24gYmVuY2hCaXRzZXQoKSB7CiAgLy8gd2FybXVwCiAgZm9yIChsZXQgaSA9IDA7IGkgPCA0OyBpKyspIHsKICAgIGJlbmNoQWRkKCk7CiAgICBiZW5jaEhhcygpOwogICAgY2xlYXIoKTsKICB9CgogIGxldCB0ID0gcGVyZm9ybWFuY2Uubm93KCk7CiAgYmVuY2hBZGQoKTsKICB0ID0gcGVyZm9ybWFuY2Uubm93KCkgLSB0OwogIG91dHB1dC52YWx1ZSArPSBgQml0U2V0IGFkZDogJHt0LnRvRml4ZWQoNCl9XG5gCgogIHQgPSBwZXJmb3JtYW5jZS5ub3coKTsKICBiZW5jaEhhcygpOwogIHQgPSBwZXJmb3JtYW5jZS5ub3coKSAtIHQ7CiAgb3V0cHV0LnZhbHVlICs9IGBCaXRTZXQgaGFzOiAke3QudG9GaXhlZCg0KX1cblxuYAp9CgpmdW5jdGlvbiBiZW5jaFNldCgpIHsKICAvLyB3YXJtdXAKICBmb3IgKGxldCBpID0gMDsgaSA8IDQ7IGkrKykgewogICAgYmVuY2hTZXRBZGQoKTsKICAgIGJlbmNoU2V0SGFzKCk7CiAgICBjbGVhclNldCgpOwogIH0KCiAgbGV0IHQgPSBwZXJmb3JtYW5jZS5ub3coKTsKICBiZW5jaFNldEFkZCgpOwogIHQgPSBwZXJmb3JtYW5jZS5ub3coKSAtIHQ7CiAgb3V0cHV0LnZhbHVlICs9IGBTZXQ8aTMyPiBhZGQ6ICR7dC50b0ZpeGVkKDQpfVxuYAoKICB0ID0gcGVyZm9ybWFuY2Uubm93KCk7CiAgYmVuY2hTZXRIYXMoKTsKICB0ID0gcGVyZm9ybWFuY2Uubm93KCkgLSB0OwogIG91dHB1dC52YWx1ZSArPSBgU2V0PGkzMj4gaGFzOiAke3QudG9GaXhlZCg0KX1cbmAKfQoKCmJlbmNoQml0c2V0KCk7CmJlbmNoU2V0KCk7CgoKCjwvc2NyaXB0Pg==)

My results for Chrome:
```
BitSet add: 39.5000
BitSet has: 29.3000

Set<i32> add: 194.6000
Set<i32> has: 207.3000
```