@HerrCai0907 2022-07-10
This algorithm use 1bit to store whether i32 number exists. What will happen if adding I32.MAX? It looks like will alloc 536870912(= 2 ** 32 / 8) bytes memory.