@MaxGraey 2022-07-10
> This algorithm use 1bit to store whether i32 number exists. What will happen if adding I32.MAX? It looks like will alloc 536870912(= 2 ** 32 / 8) bytes memory.

Yes, `BetSet` should be used for low-sparse scenarios. Where it is actually used. The maximum number of locales per-function is 65536 for most Wasm engines. In fact, in real apps locale count does not exceed 50 per function (even after inlines). This will often be within backed buffer size